Aktualne počasie na jedno miesto
Chybná teplota neviem prečo :)

+++ src/main/java/com/example/luk/weather/Coord.java
@@ -1,26 +1,26 @@
 package com.example.luk.weather;
 
 public class Coord {
-    private int lon;
-    private int lat;
+    private double lon;
+    private double lat;
 
     public Coord() {
 
     }
 
-    public int getLon() {
+    public double getLon() {
         return lon;
     }
 
-    public void setLon(int lon) {
+    public void setLon(double lon) {
         this.lon = lon;
     }
 
-    public int getLat() {
+    public double getLat() {
         return lat;
     }
 
-    public void setLat(int lat) {
+    public void setLat(double lat) {
         this.lat = lat;
     }
